+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Can set default portal from an env. Call `koop/agol/_env` to access it.
+
 ## [3.6.4] - 2016-05-10
 ### Fixed
 * Better error handling for csv errors